Refactor `SwapTransactionConfirmationView` into `SaferSignTransactionContainerView` & `SaferSignTransactionView`, remove dependency on `ParsedTransaction` for re-use by Sign panel.
